% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiostringsProjectNEW.R
\name{align}
\alias{align}
\title{Return GRanges of uniquely mapping hits.}
\usage{
align(views, genome = NULL, ...)
}
\arguments{
\item{views}{The \code{\link[BSgenome]{BSgenomeViews}} DNA to mapped.}

\item{genome}{The \code{\link[BSgenome]{BSgenome}} DNA to search for hits.}

\item{...}{Extra arguments passed on to \code{\link[QuasR]{qAlign}}.}
}
\value{
The \code{\link[GenomicRanges]{GRanges-class}} of uniquely mapping DNA sequences.
}
\description{
Return GRanges of uniquely mapping hits.
}
