% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappable.R
\name{expand_rle}
\alias{expand_rle}
\title{Insert into Rle runLength.}
\usage{
expand_rle(rle, size)
}
\arguments{
\item{rle}{Rle object to expand}

\item{size}{Rle with runValues of 1 for no insertion and runValues of > 1
corresponding to insertions.  The runLength encodes the positions of
insertions in \code{rle}}
}
\value{
Rle with insertions
}
\description{
Insert into Rle runLength.
}
\examples{
rle <- S4Vectors::Rle(c("a", "a", "a", "b", "c", "c"))
rle
size <- S4Vectors::Rle(c(1, 2, 1, 2, 2, 1))
kmap:::expand_rle(rle, size)
}
