% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiostringsProjectNEW.R
\name{gr_masked}
\alias{gr_masked}
\title{Remove motifs from BSgenomeViews}
\usage{
gr_masked(views, motif = "N")
}
\arguments{
\item{views}{The \code{\link[BSgenome]{BSgenomeViews}} DNA to subset.}

\item{motif}{The motif to remove in the sequences.}
}
\value{
A \code{\link[GenomicRanges]{GRanges-class}} object with \code{motif}
    sequences removed.
}
\description{
Filter out DNA motifs to create a \code{\link[GenomicRanges]{GRanges-class}}
instance.
}
