% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmap.R
\docType{package}
\name{kmap}
\alias{kmap}
\alias{kmap-package}
\title{kmap: Generate mappable regions of the genome for a given K-mer length}
\description{
kmap provides the \code{\link{mappable}} function.
}
\details{
\code{\link{mappable}} uses Bowtie to determine regions with unique mapping
reads of a given length k; in other words, a given k-mer length.

Finding mappable regions can take a few minutes even for small genomes,
therefore \code{\link{mappable}} prints informative messages about processing
steps using the \code{\link[futile.logger]{futile.logger}} package.
}
