% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiostringsProjectNEW.R
\name{mappable}
\alias{mappable}
\title{Return mappable regions.}
\usage{
mappable(genome, BPPARAM = bpparam())
}
\arguments{
\item{genome}{The \code{\link[BSgenome]{BSgenome}} DNA to be subset.}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}}
instance determining the parallel back-end to be used during evaluation,
or a \code{\link[base]{list}} of
\code{\link[BiocParallel]{BiocParallelParam}} instances, to be applied in
sequence for nested calls to \code{BiocParallel} functions.}
}
\value{
The \code{\link[GenomicRanges]{GRanges-class}} of mappable DNA sequences.
}
\description{
Return mappable regions.
}
\examples{

\dontrun{
mappable(BSgenome.Hsapiens.UCSC.hg38,
}
 
}
