% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappable.R
\name{mappable_cache}
\alias{mappable_cache}
\alias{mappable_cache_name}
\alias{mappable_cache_path}
\alias{mappable_cache_save}
\alias{mappable_cache_load}
\title{Internal helper functions to manage mappable GRanges in BiocFileCache.}
\usage{
mappable_cache_name(bsgenome, kmer = 36)

mappable_cache_path(name, cache_path = NULL)

mappable_cache_save(granges, name, cache_path = NULL, force = FALSE)

mappable_cache_load(name, bsgenome, cache_path = NULL)
}
\arguments{
\item{bsgenome}{Apply the sequence information from this
\code{\link[BSgenome]{BSgenome}} or \code{\link[BSgenome]{BSgenomeViews}}
object to loaded \code{\link[GenomicRanges]{GRanges}}.}

\item{kmer}{The exact size of DNA segments to check for unique mapping in the
genome.}

\item{name}{Reference name generated by \code{\link{mappable_cache_name}} to
access \code{\link[BiocFileCache]{BiocFileCache}}.}

\item{cache_path}{Path to initialize
\code{\link[BiocFileCache]{BiocFileCache}} object at a non-default path
storage location.}

\item{granges}{\code{\link[GenomicRanges]{GRanges}} to save to cache.}

\item{force}{Save to cache even if an object already exists.  Only used to
simulate a name collision for unit tests and not indended for use
otherwise; using this will break your ability to run
\code{\link{mappable_cache_load}}!}
}
\value{
\code{\link{mappable_cache_name}} returns a character vector
    reference name for a given bsgenome and kmer size.

\code{\link{mappable_cache_path}} reutrns a character vector of the
    path to the cached file path if the file exists, otherwise NULL.

\code{\link{mappable_cache_save}} returns the path to the saved cache
    file.

\code{\link{mappable_cache_load}} returns the
    \code{\link[GenomicRanges]{GRanges}} of mappable areas of the genome from
    the cache.
}
\description{
Reference name for mappable genome of a given k-mer size.

Path to cached GRanges.

Save GRanges to cache.

Load GRanges from cache.
}
