% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiostringsProjectNEW.R
\name{range_hits}
\alias{range_hits}
\alias{ranges_hits}
\title{Find hits of kmers along genome.}
\usage{
range_hits(xstring, pdict)

ranges_hits(views, pdict, indices = NULL)
}
\arguments{
\item{xstring}{The \code{\link[Biostrings]{XString-class}}}

\item{pdict}{The \code{\link[Biostrings]{PDict}} dictionary.}

\item{views}{The \code{\link[BSgenome]{BSgenomeViews}}}

\item{indices}{Only operate on these indices of the \code{views}}
}
\value{
A \code{\link[IRanges]{CompressedNormalIRangesList-class}} with no repeating hits.

A \code{\link[GenomicRanges]{GRanges-class}} with no repeating hits.
}
\description{
Compare \code{\link[Biostrings]{XString-class}} object with the
dictionary of the kmerized genome, then vectorize across
}
