% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappable.R
\name{timeit}
\alias{timeit}
\title{Run code with message and timing information.}
\usage{
timeit(msg, code)
}
\arguments{
\item{msg}{Message describing the code block being run.}

\item{code}{Block of code to evaluate and time.  Enclose in curly braces for
multiple lines.}
}
\value{
Invisibly returns evaluation time taken.
}
\description{
Only print messages if the magic variable \code{verbose} is set in the
environment from where \code{timeit} was called.
}
\examples{
## Use timeit directly inside a script
verbose <- TRUE
timeit("My long calculation", Sys.sleep(0.5))
## Silence the same calculation
verbose <- FALSE
sec <- timeit("My long calculation", Sys.sleep(0.5))
sec

## Use timeit inside a function
my_long_calculation <- function(verbose = TRUE) {
    timeit("Step 1", Sys.sleep(0.1))
    timeit("Step 2", {
        x <- 10
        Sys.sleep(0.2)
    })
    x
}
my_long_calculation()
my_long_calculation(verbose = FALSE)

}
