% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BiostringsProjectNEW.R
\name{ir2gr}
\alias{ir2gr}
\title{Convert IRanges back to GRanges using metadata from Views.}
\usage{
ir2gr(ranges, views)
}
\arguments{
\item{ranges}{IRanges}

\item{views}{The \code{\link[BSgenome]{BSgenomeViews}} DNA to subset.}
}
\value{
GRanges
}
\description{
We lose the \code{\link[GenomicRanges]{seqnames}} and
\code{\link[GenomeInfoDb]{seqinfo}} metadata when transforming the underlying
\code{\link[Biostrings]{DNAString}} of a
\code{\link[BSgenome]{BSgenomeViews}} object into
\code{\link[IRanges]{IRanges}}.  \code{\link{ir2gr}} recovers that
information from the original \code{\link[BSgenome]{BSgenomeViews}} to up
convert \code{\link[IRanges]{IRanges}} to
\code{\link[GenomicRanges]{GRanges}}.
}
\examples{
bsgenome <- BSgenome::getBSgenome("BSgenome.Ecoli.NCBI.20080805")
views <- BSgenomeViews(bsgenome, GRanges("AC_000091:11-20"))
views
remaining <- mask(views[[1]], "T")
ir <- as(remaining, "IRanges")
ir
## Convert IRanges to GRanges
ir2gr(ir, views)
}
