% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappable.R
\name{kmerize}
\alias{kmerize}
\title{Chop up BSgenomeViews into overlapping k-mers.}
\usage{
kmerize(views, kmer = 36)
}
\arguments{
\item{views}{The \code{\link[BSgenome]{BSgenomeViews}} sequences to segment
into overlapping k-mers.}

\item{kmer}{The size of overlapping segments.}
}
\value{
A \code{\link[BSgenome]{BSgenomeViews}} object with
    \code{\link[IRanges]{width}} exactly equal to \code{kmer} size.
    Ranges smaller than the \code{kmer} are dropped.
}
\description{
Chop up BSgenomeViews into overlapping k-mers.
}
