% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mappable.R
\name{mappable}
\alias{mappable}
\title{Return mappable regions of a genome.}
\usage{
mappable(genome, kmer = 36, BPPARAM = bpparam(), verbose = TRUE,
  cache_path = NULL)
}
\arguments{
\item{genome}{The \code{\link[BSgenome]{BSgenome}} DNA to be subset.}

\item{kmer}{The size of overlapping segments.}

\item{BPPARAM}{An optional \code{\link[BiocParallel]{BiocParallelParam}}
instance determining the parallel back-end to be used during evaluation,
or a \code{\link[base]{list}} of
\code{\link[BiocParallel]{BiocParallelParam}} instances, to be applied in
sequence for nested calls to \code{BiocParallel} functions.}

\item{verbose}{Print the calculation steps and their elapsed time.}

\item{cache_path}{Path to initialize
\code{\link[BiocFileCache]{BiocFileCache}} object at a non-default path
storage location.}
}
\value{
The \code{\link[GenomicRanges]{GRanges-class}} of mappable DNA sequences.
}
\description{
Return mappable regions of a genome.
}
\examples{
\dontrun{
mappable("BSgenome.Scerevisiae.UCSC.sacCer2")
}
 
}
